# Avoid using Buffer.concat() on single part messages and provide length otherwise