Avoiding a copy means improved performance, especially with large messages.