@JoshuaWise 2016-09-17
Nice! Could you run a benchmark before and after so we can see the difference? Surely it will be faster, but I'd like to see by how much. After that I'll merge.