@Nibbler999 2016-09-17
Before:
```
  Receiver masked text message
  219500.07 ops/sec, 16846 times executed, benchmark took 5.021 sec.

  Receiver binary data (125 bytes)
  240890.52 ops/sec, 18462 times executed, benchmark took 5.046 sec.

  Receiver binary data (65535 bytes)
  34372.50 ops/sec, 2853 times executed, benchmark took 5.118 sec.

  Receiver binary data (200 kB)
  12879.60 ops/sec, 1017 times executed, benchmark took 5.107 sec.
```

After:
```
  Receiver masked text message
  228183.72 ops/sec, 17518 times executed, benchmark took 5.04 sec.

  Receiver binary data (125 bytes)
  262059.70 ops/sec, 19966 times executed, benchmark took 5.087 sec.

  Receiver binary data (65535 bytes)
  46598.28 ops/sec, 3645 times executed, benchmark took 5.056 sec.

  Receiver binary data (200 kB)
  17356.50 ops/sec, 1354 times executed, benchmark took 5.076 sec.
``